+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-* references in LaTeX IGT examples
+* references in LaTeX and HTML IGT examples
 * proper handling of glossing abbreviations in LaTeX output
 
 ### Removed