+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
+
+### Added
+
+### Removed
+
+### Changed
+
+
+## [0.0.3] -- 2022-05-27
 
 ### Added
 * manual (i.e., non-interlinear) examples (`[manex](id)`)
@@ -58,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[Unreleased]: https://github.com/fmatter/pylingdocs/compare/0.0.2...HEAD
+[Unreleased]: https://github.com/fmatter/pylingdocs/compare/0.0.3...HEAD
+[0.0.3]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.3
 [0.0.2]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.2
 [0.0.1]: https://github.com/fmatter/pylingdocs/releases/tag/v0.0.1